requirements.txt
fastapi
uvicorn
httpx
openai
google-generativeai
python-dotenv
sqlalchemy
pymysql
yfinance
